(feat) Implememt regionEditor
Super hacky and I hate it. But it works nice. Click a polygon and the editor
exposes a data binding to just that polygon's action data inside the
currEventData object on the editor scope. User can then edit in place, and save
the entire event object as before. This commit also changes the server to serve
static files correctly and changes client-side polygon styling to make them more
visible.