(bug) Fix multiple polygon saving issue. Closes #167.
Saving event data in the editor now also pulls the latest version of that data
from the server (after our DB generates IDs for new regions). We then set that
data on the dashboard factory as before, but the factory now emits an event in
the process, which is heard by the editor controller which causes it to refresh
the data on its scope with the latest version from the factory.